refactor(frontend): change RadioInputOption file name bb-134
