+++ apps/frontend/src/libs/types/radio-input-option.type.ts
